Moved files from icons folder to logos folder & fixed paths
