+++ src/lib/assets/logos/Arrow.svelte
